Move daemon network related code into own module
