+++ daemon/network/response_helper.rs
